requirements.txt
boto3==1.9.60
Cerberus==1.2
docx-mailmerge==0.4.0
Flask==1.0.2
